Do a check on the file path passed to indykey
This is a security fix. Previously, indykey would allow unauthenticated access to any
file on the filesystem, with the potential to overwrite its contents and write relative
junk to it. This changes this behaviour to check the path passed against an internally-
hardcoded list of acceptable paths and refuse to operate on any path which isn't
present in the list. This can be circumvented by importing/running indykey from a
root shell/interpreter, which takes care of the authentication for us.

This approach is secure because the indykey source code is required to be installed
as root, due to the need for the policykit access files to be installed. A user-mode
installation of indykey will not function. Therefore, the hardcoded list is protected
by filesystem permissions that prevent its modification by non-root users.